Gameplay Update 10
Combined UI and Gameplay updates into one scene

added lock-on bullet (graphic temporary)
Changed bullet penetration slightly, gave them death animations
Fixed issue with script crashing if there was no defined death animation

Lock on bullet: exits launcher slowly, then finds a target and homes in
on them.  It explodes in a small AoE on impact, or selects a new target
if its target dies before it hits anything.  Currently used on Bass

Undocumented previous change: Added icon to executable

+++ project/Assets/Code/Gameplay/Bullets/CircularCurveBullet.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class CircularCurveBullet : ShotObject {
 
-	public int rotationSpeed = 180; //how many degrees to turn in one second
+	public float rotationSpeed = 180f; //how many degrees to turn in one second
 	
 	// Rotate and move forward (moves in a circle)
 	new void Update () {
